requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-ninja==1.4.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
frozenlist==1.6.0
idna==3.10
kombu==5.5.3
MarkupSafe==3.0.2
multidict==6.4.4
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
redis==6.2.0
requests==2.32.3
sendgrid==6.12.3
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.0
